Export redundant code to Abstract parser class
